refactor: 패널 폴더 구조 변경
